Final generate links

+++ generateLinks.php
@@ -1,9 +1,11 @@
 <?php
 include('init.php');
 set_time_limit(3600);
-mysql_query("TRUNCATE wg_links");
-// mysql_query("UPDATE wg_page SET visited=0");
-$r = mysql_query("SELECT * FROM wg_page WHERE visited=0 ORDER BY id LIMIT 3");
+if(isset($_GET['fullReset'])) {
+	mysql_query("TRUNCATE wg_links");
+	mysql_query("UPDATE wg_page SET visited=0");
+}
+$r = mysql_query("SELECT * FROM wg_page WHERE visited=0 ORDER BY id");
 while($re = mysql_fetch_array($r)) {
 	$html = file_get_contents('data/'.$re['id'].'.txt');
 	$dom = new DOMDocument;
